[fix] razorpay merged in integrations
